+++ CMakeLists.txt
@@ -87,4 +87,4 @@ endif()
 add_subdirectory(test)
 add_subdirectory(tools)
 enable_testing()
-#add_subdirectory(unittest)
+add_subdirectory(unittest)
