+++ lib/Dialect/Quantum/Transforms/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_mlir_dialect_library(QuantumTransforms
-        Optimise.cpp
+        Hermitian.cpp
+        GateOptimization.cpp
         MultiQubitLegalization.cpp
         ScfToRVSDG.cpp
