+++ lib/Dialect/Quantum/Transforms/GateOptimization.cpp
@@ -1,0 +1,77 @@
+/// Implements the gate optimization.
+///
+/// @file
+/// @author     Lars Schütze (lars.schuetze@tu-dresden.de)
+/// @author     Washim Neupane (washimneupane@outlook.com)
+
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/PatternMatch.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+#include "quantum-mlir/Dialect/Quantum/IR/Quantum.h"
+
+using namespace mlir;
+using namespace mlir::quantum;
+
+//===- Generated includes -------------------------------------------------===//
+
+namespace mlir::quantum {
+
+#define GEN_PASS_DEF_QUANTUMOPTIMISE
+#include "quantum-mlir/Dialect/Quantum/Transforms/Passes.h.inc"
+
+} // namespace mlir::quantum
+
+//===----------------------------------------------------------------------===//
+
+namespace {
+
+struct QuantumOptimisePass
+        : mlir::quantum::impl::QuantumOptimiseBase<QuantumOptimisePass> {
+    using QuantumOptimiseBase::QuantumOptimiseBase;
+
+    void runOnOperation() override;
+};
+
+/// Pattern: Drop phase gates immediately before measurement
+struct DropPhaseBeforeMeasure : OpRewritePattern<quantum::ZOp> {
+    using OpRewritePattern<quantum::ZOp>::OpRewritePattern;
+
+    LogicalResult
+    matchAndRewrite(ZOp zOp, PatternRewriter &rewriter) const override
+    {
+        Value zResult = zOp.getResult();
+        if (!zResult.getType().isa<quantum::QubitType>()) return failure();
+        if (zResult.use_empty()) return failure();
+
+        for (Operation* userOp : zResult.getUsers()) {
+            auto measureOp = dyn_cast<MeasureOp>(userOp);
+            if (!measureOp) return failure();
+            // Ensure that the qubit operand of the MeasureOp is exactly
+            // zResult
+            if (measureOp.getOperand() != zResult) return failure();
+        }
+        rewriter.replaceOp(zOp, zOp.getInput());
+        return success();
+    }
+};
+
+} // namespace
+
+void QuantumOptimisePass::runOnOperation()
+{
+    RewritePatternSet patterns(&getContext());
+    populateQuantumOptimisePatterns(patterns);
+    if (failed(applyPatternsGreedily(getOperation(), std::move(patterns))))
+        signalPassFailure();
+}
+
+void mlir::quantum::populateQuantumOptimisePatterns(RewritePatternSet &patterns)
+{
+    patterns.add<DropPhaseBeforeMeasure>(patterns.getContext());
+}
+
+std::unique_ptr<Pass> mlir::quantum::createQuantumOptimisePass()
+{
+    return std::make_unique<QuantumOptimisePass>();
+}
